[REVROBO-7] Refactor file names and locations
