+++ src/containers/HelpTab.tsx
@@ -1,6 +1,6 @@
 import {Alert, Button, Dialog, ProgressBar} from "@blueprintjs/core";
 import * as React from "react";
-import {IApplicationState} from "../store/types";
+import {IApplicationState} from "../store/state";
 import {connect} from "react-redux";
 import SparkManager from "../managers/SparkManager";
